move files around; split up module
