+++ example-app/app/transfer/Gas.tsx
@@ -3,12 +3,16 @@ import { formatGwei } from 'viem'
 import { useGasPrice } from 'wagmi'
 
 export const Gas = () => {
-  const gas = useGasPrice()
+  const { isFetched, data, error } = useGasPrice()
+
+  if (error != null) {
+    return null
+  }
 
   return (
     <div className="flex items-center gap-2 text-xs">
       <Fuel size={14} />
-      {gas.isFetched && formatGwei(gas.data)}
+      {isFetched && formatGwei(data)}
     </div>
   )
 }
