+++ example-app/app/transfer/balance/BalanceValue.ts
@@ -1,0 +1,1 @@
+export type BalanceValue = [balance: string, symbol: string]
