+++ example-app/app/transfer/balance/EthBalance.tsx
@@ -1,0 +1,30 @@
+import { Address, formatEther } from 'viem'
+import { useBalance } from 'wagmi'
+import type { BalanceValue } from './BalanceValue'
+import { Symbol } from './Symbol'
+
+type EthBalanceProps = {
+  address: Address
+}
+
+export const EthBalance = ({ address }: EthBalanceProps) => {
+  const [balance, symbol] = useEthBalance(address)
+
+  return (
+    <>
+      {balance}
+
+      <Symbol>{symbol}</Symbol>
+    </>
+  )
+}
+
+const useEthBalance = (address: Address): BalanceValue | [null, null] => {
+  const { data, isFetched } = useBalance({ address })
+
+  if (isFetched) {
+    return [formatEther(data.value), data.symbol] as const
+  }
+
+  return [null, null]
+}
