+++ example-app/app/transfer/balance/index.ts
@@ -1,0 +1,1 @@
+export { Balance } from './Balance'
